+++ CHANGELOG.md
@@ -362,6 +362,7 @@
 - [9911](https://github.com/vegaprotocol/vega/issues/9911) - Bind correct property in `GraphQL` resolver.
 - [9913](https://github.com/vegaprotocol/vega/issues/9913) - Add missing properties in schema.
 - [9916](https://github.com/vegaprotocol/vega/issues/9916) - Fix activity streak locked account until epoch.
+- [9924](https://github.com/vegaprotocol/vega/issues/9924) - Referral set referees `API` should aggregate by number of epochs.
 
 ## 0.72.1